Switch to a single higher-order function
